try zu

+++ hello_world/6-size.c
@@ -11,11 +11,11 @@ int main(void)
 	long long ll;
 	float f;
 
-	printf("Size of a char: %lu\n", (unsigned long)sizeof(c));
-	printf("Size of an int: %lu\n", (unsigned long)sizeof(i));
-	printf("Size of a long int: %lu\n", (unsigned long)sizeof(l));
-	printf("Size of a long long int: %lu\n", (unsigned long)sizeof(ll));
-	printf("Size of a float: %lu\n", (unsigned long)sizeof(f));
+	printf("Size of a char: %zu\n", sizeof(c));
+	printf("Size of an int: %zu\n", sizeof(i));
+	printf("Size of a long int: %zu\n", sizeof(l));
+	printf("Size of a long long int: %zu\n", sizeof(ll));
+	printf("Size of a float: %zu\n", sizeof(f));
 
 	return (0);
 }
